+++ src/fields/p192.h
@@ -42,7 +42,7 @@ p192_fe_equal(const p192_fe_t x, const p192_fe_t y) {
   uint32_t z = 0;
   uint8_t u[24];
   uint8_t v[24];
-  size_t i;
+  int i;
 
   fiat_p192_to_bytes(u, x);
   fiat_p192_to_bytes(v, y);
@@ -54,8 +54,8 @@ p192_fe_equal(const p192_fe_t x, const p192_fe_t y) {
 }
 
 static void
-p192_fe_sqrn(p192_fe_t r, const p192_fe_t x, unsigned int n) {
-  unsigned int i;
+p192_fe_sqrn(p192_fe_t r, const p192_fe_t x, int n) {
+  int i;
 
   p192_fe_sqr(r, x);
